backend/requirements.txt
flask
flask-cors
opencv-python
pyserial
numpy
